Plot and evalind() fix
Changed code that could cause error in plotting with a very small number
of variables in the data set.

Fixed code in the individual-level pruning stage that could cause
significant paths with negative coefficients to be erroneously removed.
Paths were frequently added back in the fixfit() function, but not
always. Group paths, subgroup results, and subgroup paths unaffected by
this bug.