Roll back change to z_list
to kmg - I'm rolling back this change. It will likely have unintended consequences. At the time of subgrouping, all persons have the same number of rows, which is important for the creation of the similarity matrix. If you exclude the rows with NAs, there will be off-by-one errors, and the similarity matrix won't be meaningful. My recent change adding na.rm = T should address the cases you want.