Created sparsity.optimization script
Script was originally embedded within the determine.subgroups script but was taking up a lot of space.

Takes the generated similarity matrix and the sub_method argument to numerically optimize modularity within a given sub_method. The resulting value is used as a percentile with which all values <= are reduced to 0.00 to induce sparsity.

Credit to Zack for coding the optimization.